Extract useMenuPosition hook from Dropdown
Deduplicate viewport-aware positioning logic shared between
Dropdown and FilterMenu into a reusable hook.
